Moved components into subfolder for better organization. Updated imports.
