+++ src/applications/representative-search/components/mapbox/MapboxClient.js
@@ -1,5 +1,5 @@
 import MapboxClient from '@mapbox/mapbox-sdk';
-import { mapboxToken } from '../utils/mapboxToken';
+import { mapboxToken } from '../../utils/mapboxToken';
 
 const mapboxClient = new MapboxClient({ accessToken: mapboxToken });
 export default mapboxClient;
