+++ src/applications/representative-search/components/results/SearchResult.jsx
