+++ src/applications/representative-search/containers/App.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import environment from '@department-of-veterans-affairs/platform-utilities/environment';
 import { VaLoadingIndicator } from '@department-of-veterans-affairs/component-library/dist/react-bindings';
-import GetFormHelp from '../components/GetFormHelp';
+import GetFormHelp from '../components/footer/GetFormHelp';
 import { useFeatureToggle } from '~/platform/utilities/feature-toggles/useFeatureToggle';
 
 // import DowntimeNotification, {
