+++ src/applications/representative-search/tests/components/SearchResultsHeader.unit.spec.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { expect } from 'chai';
 import { shallow } from 'enzyme';
-import { SearchResultsHeader } from '../../components/search/SearchResultsHeader';
+import { SearchResultsHeader } from '../../components/results/SearchResultsHeader';
 // import { RepresentativeType } from '../../constants';
 import testDataRepresentative from '../../constants/mock-representative-v0.json';
 import testDataResponse from '../../constants/mock-representative-data.json';
