+++ src/applications/representative-search/utils/mapHelpers.js
@@ -1,5 +1,5 @@
 import mbxGeo from '@mapbox/mapbox-sdk/services/geocoding';
-import mapboxClient from '../components/MapboxClient';
+import mapboxClient from '../components/mapbox/MapboxClient';
 
 import { BOUNDING_RADIUS } from '../constants';
